refactor(dtx-web): move download.ts to lib/api, dtx-api-only
Phase 6 — the download helpers now always fetch from dtx-api
(/downloads/*) with a bearer header. Drop the local-REST direct-link
branch and relocate out of the misnamed rest/ dir.
