+++ packages/dtx-web/src/lib/api/index.ts
@@ -2,4 +2,4 @@ export * from './chart';
 export * from './user';
 export * from './auth';
 export * from './client';
-export { downloadSimfile, bulkDownloadBaseUrl, bulkDownloadHeaders } from './rest/download';
+export { downloadSimfile, bulkDownloadBaseUrl, bulkDownloadHeaders } from './download';
